+++ requirements-dev.txt
@@ -3,4 +3,3 @@ identify>=1.4.20
 mypy>=0.931
 pre-commit
 pytest>=6.2.5
-types-PyYAML>=6.0.3
