--- src/template/example.py
@@ -1,9 +1,0 @@
-import numpy as np  # type: ignore[import]
-
-
-def arange(number):
-    return np.arange(number)
-
-
-def add_one(number):
-    return number + 1
