+++ src/template/sanity/__init__.py
@@ -1,0 +1,1 @@
+from template.sanity import example
