+++ src/template_duplicate/sanity/__init__.py
@@ -1,0 +1,1 @@
+from template_duplicate.sanity import example
